requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
idna==2.8
Pillow==5.3.0
pycparser==2.19
six==1.12.0
